refactor: initializer interface refactored to be more simple
- remove initial data
- add BuildingInitializer injection
