+++ Assets/_Scripts/Buildings/Initializing/BuildingInitializerInstaller.cs
@@ -1,0 +1,10 @@
+using UnityEngine;
+using Zenject;
+
+public class BuildingInitializerInstaller : MonoInstaller
+{
+    public override void InstallBindings()
+    {
+        Container.BindInstance(GetComponent<IInitializer<Building>>());
+    }
+}
