+++ Assets/_Scripts/General/IObserver.cs
@@ -1,3 +1,7 @@
+/// <summary>
+/// Basic class for implementing Observer pattern logic.
+/// </summary>
+/// <typeparam name="T">Type of subscribers objects.</typeparam>
 public interface IObserver<T>
 {
     void AddSubscriber(T subscriber);
